Latest stable changes from develop

+++ src/backend/src/main/java/com/linkedpipes/lpa/backend/services/UserServiceComponent.java
@@ -5,6 +5,8 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import com.linkedpipes.lpa.backend.entities.*;
 import com.linkedpipes.lpa.backend.entities.profile.*;
 import com.linkedpipes.lpa.backend.entities.database.*;
+import com.linkedpipes.lpa.backend.services.virtuoso.VirtuosoService;
+import com.linkedpipes.lpa.backend.exceptions.LpAppsException;
 import com.linkedpipes.lpa.backend.exceptions.UserNotFoundException;
 import com.linkedpipes.lpa.backend.util.LpAppsObjectMapper;
 import org.jetbrains.annotations.NotNull;
@@ -13,6 +15,7 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Profile;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.transaction.annotation.Isolation;
@@ -44,6 +47,9 @@ public class UserServiceComponent implements UserService {
 
     @Autowired
     private DiscoveryNamedGraphRepository ngRepository;
+
+    @Autowired
+    private ApplicationRepository appRepository;
 
     /**
     * Returns the user's profile. If user doesn't exist yet we add them.
@@ -197,18 +203,17 @@ public class UserServiceComponent implements UserService {
         profile.pipelineExecutions = new ArrayList<>();
         if (user.getExecutions() != null) {
             for (ExecutionDao e : user.getExecutions()) {
-                PipelineExecution exec = new PipelineExecution();
-                exec.status = e.getStatus();
-                exec.executionIri = e.getExecutionIri();
-                exec.etlPipelineIri = e.getPipeline().getEtlPipelineIri();
-                exec.selectedVisualiser = e.getSelectedVisualiser();
-                exec.started = e.getStarted().getTime() / 1000L;
-                if (e.getFinished() != null) {
-                    exec.finished = e.getFinished().getTime() / 1000L;
-                } else {
-                    exec.finished = -1;
-                }
-                profile.pipelineExecutions.add(exec);
+                profile.pipelineExecutions.add(PipelineExecution.getPipelineExecutionFromDao(e));
+            }
+        }
+
+        profile.applications = new ArrayList<>();
+        if (user.getApplications() != null) {
+            for (ApplicationDao application : user.getApplications()) {
+                Application app = new Application();
+                app.solidIri = application.getSolidIri();
+                app.executionAvailable = (application.getExecution() != null) && (!application.getExecution().isRemoved());
+                profile.applications.add(app);
             }
         }
 
@@ -232,7 +237,6 @@ public class UserServiceComponent implements UserService {
                                        throws UserNotFoundException {
         UserDao user = getUser(username);
         ExecutionDao toDelete = null;
-        PipelineInformationDao pipelineInformationToDelete = null;
 
         for (ExecutionDao execution : user.getExecutions()) {
             if (execution.getExecutionIri().equals(executionIri)) {
@@ -241,6 +245,14 @@ public class UserServiceComponent implements UserService {
             }
         }
 
+        removeExecutionFromUser(user, toDelete);
+
+        return transformUserProfile(user);
+    }
+
+    private void removeExecutionFromUser(UserDao user, ExecutionDao toDelete) {
+        PipelineInformationDao pipelineInformationToDelete = null;
+
         if (toDelete != null) {
             user.removeExecution(toDelete);
 
@@ -248,16 +260,30 @@ public class UserServiceComponent implements UserService {
                 pipelineInformationToDelete = toDelete.getPipeline();
             }
 
-            executionRepository.delete(toDelete);
+            List<ExecutionDao> executions = new ArrayList<>();
+            for (PipelineInformationDao x : pipelineRepository.findByResultGraphIri(toDelete.getPipeline().getResultGraphIri())) {
+                executions.addAll(x.getExecutions());
+            }
 
-            if (pipelineInformationToDelete != null) {
-                pipelineRepository.delete(pipelineInformationToDelete);
+            if (toDelete.getApplications().isEmpty()) { //this is for repeated executions using the same NG
+                //no applications - delete from virtuoso
+                String graphName = toDelete.getPipeline().getResultGraphIri();
+                VirtuosoService.deleteNamedGraph(graphName);
+
+                for (ExecutionDao e : executions) {
+                    executionRepository.delete(e);
+                }
+
+                if (pipelineInformationToDelete != null) {
+                    pipelineRepository.delete(pipelineInformationToDelete);
+                }
+            } else {
+                toDelete.setRemoved(true);
+                executionRepository.save(toDelete);
             }
 
             repository.save(user);
         }
-
-        return transformUserProfile(user);
     }
 
     /**
@@ -308,6 +334,88 @@ public class UserServiceComponent implements UserService {
         user.setColor(color);
         repository.save(user);
         return transformUserProfile(user);
+    }
+
+    @Override @Transactional(rollbackFor=UserNotFoundException.class)
+    public UserProfile addApplication(String username, String executionIri, String solidIri) throws UserNotFoundException {
+        UserDao user = getUser(username);
+
+        for (ExecutionDao e : user.getExecutions()) {
+            if (e.getExecutionIri().equals(executionIri)) {
+                ApplicationDao app = new ApplicationDao();
+                app.setSolidIri(solidIri);
+                e.addApplication(app);
+                user.addApplication(app);
+
+                repository.save(user);
+                executionRepository.save(e);
+                appRepository.save(app);
+                break;
+            }
+        }
+
+        return transformUserProfile(user);
+    }
+
+    @Override @Transactional(rollbackFor=UserNotFoundException.class)
+    public UserProfile deleteApplication(String username, String solidIri) throws UserNotFoundException {
+        UserDao user = getUser(username);
+
+        for (ApplicationDao app : user.getApplications()) {
+            if (app.getSolidIri().equals(solidIri)) {
+                ExecutionDao execution = app.getExecution();
+                if (null != execution) {
+                    if (execution.isScheduled()) {
+                        for (PipelineInformationDao x : pipelineRepository.findByResultGraphIri(execution.getPipeline().getResultGraphIri())) {
+                            for (ExecutionDao y : x.getExecutions()) {
+                                if (!y.getExecutionIri().equals(execution.getExecutionIri())) {
+                                    executionRepository.delete(y);
+                                }
+                            }
+                        }
+                    }
+
+                    boolean allRemoved = execution.isRemoved();
+                    if (allRemoved) {
+                        for (PipelineInformationDao x : pipelineRepository.findByResultGraphIri(execution.getPipeline().getResultGraphIri())) {
+                            for (ExecutionDao y : x.getExecutions()) {
+                                if (!y.isRemoved()) {
+                                    allRemoved = false;
+                                    break;
+                                }
+                            }
+                        }
+                    }
+
+                    if (allRemoved) {
+                        //all executions were already removed
+                        String graphName = app.getExecution().getPipeline().getResultGraphIri();
+                        VirtuosoService.deleteNamedGraph(graphName);
+
+                        executionRepository.delete(execution);
+                    } else {
+                        execution.removeApplication(app);
+                        executionRepository.save(execution);
+                    }
+                }
+
+                user.removeApplication(app);
+                appRepository.delete(app);
+                repository.save(user);
+                break;
+            }
+        }
+
+        return transformUserProfile(user);
+    }
+
+
+
+    @Override
+    public PipelineExecution getExecution(@NotNull final String executionIri) throws LpAppsException {
+        List<ExecutionDao> lst = executionRepository.findByExecutionIri(executionIri);
+        if (!lst.isEmpty()) return PipelineExecution.getPipelineExecutionFromDao(lst.get(0));
+        throw new LpAppsException(HttpStatus.NOT_FOUND, "No such execution");
     }
 
 }
